Update roller.py
Getting back into Python and found your project, thought I'd get to work on some things you suggested in the readme.  Very basic menu loop and that fancy for loop with some extra stuff to get rid of the hard coding!  

And yes, it took severe procrastination to get to this point.